+++ src/test/bash/editor_interfaces_get.sh
@@ -1,0 +1,10 @@
+#!/usr/bin/env bash
+
+output="$(basename $(echo $0) | sed 's#sh#json#g' | sed 's#^#../resources/#g')"
+
+curl --verbose \
+    -X GET \
+    -H 'Content-Type: application/vnd.contentful.management.v1+json' \
+    -H 'Authorization: Bearer '$CMA_TOKEN  \
+    'https://api.contentful.com/spaces/'$SPACE_ID'/content_types/'$CONTENT_TYPE_ID'/editor_interface' \
+    | tee ${output}